move function signature to expr crate
