+++ src/expr/src/lib.rs
@@ -25,9 +25,11 @@
 #![feature(generators)]
 #![feature(iterator_try_collect)]
 #![feature(exclusive_range_pattern)]
+#![feature(once_cell)]
 
 pub mod error;
 pub mod expr;
+pub mod sig;
 pub mod table_function;
 pub mod vector_op;
